Printing method for KScens object.

Printing method for KScens object.

# S3 method for class 'KScens'
print(x, prnt = TRUE, outp = c("list", "table"),  degs = 3, print.AIC = TRUE,
      print.BIC = TRUE, print.infoBoot = FALSE, ...)

Arguments

x

An object of class KScens.

prnt

Logical to indicate if the estimations of the Kolmogorov-Smirnov statistic and p-value should be printed. Default is TRUE.

outp

Indicator of how the output will be displayed. The possible formats are list and table.

degs

Integer indicating the number of decimal places of the numeric results of the output.

print.AIC

Logical to indicate if the AIC of the model should be printed. Default is TRUE

print.BIC

Logical to indicate if the BIC of the model should be printed. Default is TRUE

print.infoBoot

Logical to indicate if the number of bootstrap samples used should be printed. Default is FALSE

...

Additional arguments.

Value

If prnt = TRUE, a list or table (if outp = "table") containing the following components:

Distribution

Null distribution.

Hypothesis

Parameters under the null hypothesis (if params0 is provided).

Test

Vector containing the value of the modified Kolmogorov-Smirnov statistic (A), the estimated p-value (p-value), the estimation of the image of the last recorded time (F(y_m)) and the last recorded time (y_m).

Estimates

Vector with the maximum likelihood estimates of the parameters of the distribution under study.

StdErrors

Vector containing the estimated standard errors.

aic

The Akaike information criterion.

bic

The so-called BIC or SBC (Schwarz Bayesian criterion).

BS

The number of bootstrap samples used. If the modified test is used, a 0 is returned.

The list is also returned invisibly.

Author

K. Langohr, M. Besalú, M. Francisco, A. Garcia, G. Gómez.

Examples

# List output
set.seed(123)
KScens(times = rweibull(100, 12, scale = 4), distr = "weibull", BS = 99)
#> Distribution: weibull 
#> 
#> KS Test results:
#>       A p-value 
#>   0.528   0.670 
#> 
#> Parameter estimates (se):
#> shape             scale             
#> 12.241 (0.938)     4 (0.034)     
#> 
#> AIC: 84.375 
#> BIC: 89.585 
#> 

# Table output
set.seed(123)
print(KScens(times = rweibull(100, 12, scale = 4), distr = "weibull", BS = 99),
      outp = "table")
#> Distribution: weibull 
#> 
#> KS Test results:
#> ------- | -------
#> Metric  | Value  
#> ------- | -------
#> A       | 0.528  
#> p-value | 0.67   
#> ------- | -------
#> 
#> Parameter estimates:
#> --------- | --------- | ---------
#> Parameter | Value     | s.e.     
#> --------- | --------- | ---------
#> shape     | 12.241    | 0.938    
#> scale     | 4         | 0.034    
#> --------- | --------- | ---------
#> 
#> AIC: 84.375 
#> BIC: 89.585 
#>